+++ src/ts/ir/index.ts
@@ -2,7 +2,7 @@ import {Constants} from "../constants";
 import {isCtrl, isFirefox} from "../util/compatibility";
 import {
     blurEvent,
-    copyEvent, cutEvent,
+    copyEvent, cutEvent, dblclickEvent,
     dropEvent,
     focusEvent,
     hotkeyEvent,
@@ -40,6 +40,7 @@ class IR {
         this.bindEvent(vditor);
 
         focusEvent(vditor, this.element);
+        dblclickEvent(vditor, this.element);
         blurEvent(vditor, this.element);
         hotkeyEvent(vditor, this.element);
         selectEvent(vditor, this.element);